проработка интерфейса

+++ teachers.php
@@ -10,7 +10,9 @@
 	<?php
 	require_once 'php/login.php';
 	require_once 'php/functions.php';
+
 	require_once 'php/classes.php';
+
 	$conn = new mysqli($hn, $un, $ps, $db);
 	if (!$conn) die('ERROR');
 
@@ -27,9 +29,35 @@
 			echo 'Введите данные';
 		}
 	}
-	if (isset($_POST['id_to_delete']) && isset($_POST['delete_record'])){
+
+	if (isset($_POST['title'])){
+		if ($_POST['title'] != ''){
+			$title = defuse($conn, $_POST['title']);
+			$query = "INSERT INTO lessons VALUES (NULL, '$title');";
+			$result = $conn->query($query);
+			if (!$result) {echo 'Ошибка ввода';}
+			else echo 'Успешно';
+		} else {
+			echo 'Введите данные';
+		}
+	}
+
+	if (isset($_POST['id_to_delete']) && isset($_POST['delete_record']) && isset($_POST['type'])){
+		$type = $_POST['type'];
 		$id = $_POST['id_to_delete'];
-		$query = "DELETE FROM teachers WHERE id = '$id'";
+		switch ($type){
+			case 'teacher':
+				$query = "DELETE FROM teachers WHERE id = '$id'";
+				break;
+			case 'lesson':
+				$query = "DELETE FROM lessons WHERE id = '$id'";
+				break;
+			case 'current_timetable':
+				$query = "DELETE FROM lesson WHERE id = '$id'";
+				break;
+			default:
+				break;
+		}
 		$result = $conn->query($query);
 		if (!$result){
 			echo 'Ошибка при удалении';
@@ -48,23 +76,46 @@
 	</form>
 
 	<?php
-	$query = ('SELECT * FROM teachers');
-	$info = $conn->query($query);
-	$teachers = array();
-	for ($i = 0; $i < $info->num_rows; $i++){
-		$temp = $info->fetch_array(MYSQLI_ASSOC);
-		array_push($teachers, new teacher_card($conn, $temp));
-		$teachers[$i]->print_all();
-		$id = $teachers[$i]->get_id();
-
-		echo <<<_END
-		<form action='teachers.php' method='post'>
-			<input type='hidden' name='id_to_delete' value='$id'>
-			<input type='hidden' name='delete_record' value='yes'>
-			<input type='submit' value='удалить'>
+	$teacher_field = new field('teachers', $conn, __FILE__);
+	echo '<br><br>';
+	?>
+	<form action='teachers.php' method='post'>
+		<input type='text' name='title'>
+		<input type='submit' value='Отправить'>
+	</form>
+	<?php
+	$lessons_field = new field('lessons', $conn, __FILE__);
+	?>
+	<?php
+	$teachers = $conn->query('SELECT * FROM teachers');
+	$lessons = $conn->query('SELECT * FROM lessons');
+	echo "<form action='teachers.php' method='post'>";
+			echo "<select name='lesson_id'>";
+			for ($i = 0; $i < $lessons->num_rows; $i++){
+				$temp = $lessons->fetch_array(MYSQLI_ASSOC);
+				$id = $temp['id'];
+				$title = $temp['title'];
+				echo "<option value=$id>$title</option>";
+			}
+			echo "</select>";
+			echo "<select name='teacher_id'>";
+			for ($i = 0; $i < $teachers->num_rows; $i++){
+				$temp = $teachers->fetch_array(MYSQLI_ASSOC);
+				$id = $temp['id'];
+				$name = $temp['sername'] . ' ' . $temp['name'] . ' ' . $temp['patronymic'];
+				echo "<option value=$id>$name</option>";
+			}
+			echo "</select>";
+			echo <<<_END
+			<input type='text' name='classroom'>
+			<input type='text' name='weekday'>
+			<input type='text' name='position'>
+			<input type='submit' value='Отправить'>
 		</form>
 		_END;
-	}
+	?>
+	<?php
+	$lessons_field = new field('timetable', $conn, __FILE__);
 	?>
 
 </body>
